Backend/requirements.txt
absl-py==2.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
asyncpg==0.29.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
billiard==4.2.1
celery==5.3.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.104.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.3
httptools==0.6.4
huggingface-hub==0.20.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kombu==5.5.4
libclang==18.1.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.4
namex==0.1.0
networkx==3.5
nltk==3.8.1
numpy==1.26.4
openai==0.28.1
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
passlib==1.7.4
pgvector==0.2.4
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
regex==2024.11.6
requests==2.31.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.66.1
transformers==4.35.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
